requirements.txt
# Production-Optimized Dependencies for Vercel Deployment
# Minimized to reduce deployment size

# Environment variables
python-dotenv>=1.0.0

# Core ML and RL (lightweight)
numpy>=1.26.0,<2.0.0  # Python 3.12 compatible with pre-built wheels

# Database connections (essential)
redis>=4.0.0,<5.0.0
pymongo>=4.0.0,<5.0.0
certifi>=2023.0.0
dnspython>=2.0.0

# Web framework (essential)
fastapi>=0.100.0,<0.120.0
uvicorn[standard]>=0.20.0,<0.30.0
pydantic>=2.0.0,<3.0.0
httpx>=0.24.0  # For YouTube API clients

# YouTube API (essential)
google-api-python-client>=2.80.0,<3.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# OpenAI for agents
openai>=2.7.1,<3.0.0
openai-agents>=0.4.2

# REMOVED for production to reduce size:
# - scipy (heavy, not critical)
# - pandas (heavy, can be lazy loaded if needed)
# - matplotlib (not needed in API)
# - seaborn (not needed in API)
# - motor (use pymongo instead)
# - aioredis (use redis instead)
# - asyncio-mqtt (not needed)
# - Testing tools (pytest, etc.)
# - Development tools (black, flake8, mypy)
